requirements/local.txt
-r ./base.txt

Werkzeug==1.0.0 # https://github.com/pallets/werkzeug
ipdb==0.13.2  # https://github.com/gotcha/ipdb
Sphinx==2.4.4  # https://github.com/sphinx-doc/sphinx
psycopg2==2.8.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==0.770  # https://github.com/python/mypy
django-stubs==1.5.0  # https://github.com/typeddjango/django-stubs
pytest==5.3.5  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8==3.7.9  # https://github.com/PyCQA/flake8
flake8-isort==2.9.0  # https://github.com/gforcada/flake8-isort
coverage==5.0.4  # https://github.com/nedbat/coveragepy
black==19.10b0  # https://github.com/ambv/black
pylint-django==2.0.14  # https://github.com/PyCQA/pylint-django
pre-commit==2.2.0  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==2.12.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==2.2  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.2.8  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.8.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.8.0  # https://github.com/pytest-dev/pytest-django

djangorestframework==3.11.0
